refactor code into helper functions
